feat: adjustment compositions
